fixing re-renders
